Move mage targets to libraries so the shipper can use them
